Move build directory outside of server folder
