+++ lib/constants.js
@@ -12,3 +12,5 @@ export const BLOCKED_PAGES = [
   '/_app',
   '/_error'
 ]
+export const IS_BUNDLED_PAGE = /^bundles[/\\]pages.*\.js$/
+export const MATCH_ROUTE_NAME = /^bundles[/\\]pages[/\\](.*)\.js$/
