+++ server/on-demand-entry-handler.js
@@ -3,11 +3,11 @@ import { EventEmitter } from 'events'
 import { join } from 'path'
 import { parse } from 'url'
 import touch from 'touch'
-import promisify from './lib/promisify'
+import promisify from '../lib/promisify'
 import globModule from 'glob'
 import {normalizePagePath, pageNotFoundError} from './require'
-import {createEntry} from './build/webpack/utils'
-import { MATCH_ROUTE_NAME, IS_BUNDLED_PAGE } from './utils'
+import {createEntry} from '../build/webpack/utils'
+import { MATCH_ROUTE_NAME, IS_BUNDLED_PAGE } from '../lib/constants'
 
 const ADDED = Symbol('added')
 const BUILDING = Symbol('building')
