+++ server/utils.js
@@ -1,9 +1,6 @@
 import { join } from 'path'
 import { readdirSync, existsSync } from 'fs'
 
-export const IS_BUNDLED_PAGE = /^bundles[/\\]pages.*\.js$/
-export const MATCH_ROUTE_NAME = /^bundles[/\\]pages[/\\](.*)\.js$/
-
 export function getAvailableChunks (distDir) {
   const chunksDir = join(distDir, 'chunks')
   if (!existsSync(chunksDir)) return {}
